specgap/requirements.txt
# ===== Core Framework =====
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0

# ===== AI & LLM =====
google-generativeai>=0.8.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0

# ===== Document Processing =====
pdfplumber>=0.10.3
python-docx>=1.1.0
pillow>=10.2.0

# ===== OCR Support (Optional - for scanned PDFs) =====
pytesseract>=0.3.10
pdf2image>=1.16.3

# ===== Database =====
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# ===== Data Processing =====
pandas>=2.2.0
tabulate>=0.9.0

# ===== Environment & Config =====
python-dotenv>=1.0.1

# ===== Development (Optional) =====
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# httpx>=0.27.0
# black>=24.0.0
# ruff>=0.2.0
